fuzzer: prevent unwanted integer promotion in r32le
Fixes an undefined shift in
clusterfuzz-testcase-minimized-dav1d_fuzzer-5685383740784640. Credits to
oss-fuzz.

+++ tests/libfuzzer/dav1d_fuzzer.c
@@ -33,7 +33,7 @@
 #include <dav1d/dav1d.h>
 
 static unsigned r32le(const uint8_t *const p) {
-    return (p[3] << 24U) | (p[2] << 16U) | (p[1] << 8U) | p[0];
+    return ((uint32_t)p[3] << 24U) | (p[2] << 16U) | (p[1] << 8U) | p[0];
 }
 
 // expects ivf input
